feat(on): on callbacks receive state() & are waited for
setState is used synchronously on attach() and unattach(), and waited for asynchronously on
handler(), which allows to keep attach() and unattach() synchronous while still waiting for
callbacks to terminate (promise is accumulated).

BREAKING CHANGE: on() now receives state() for convenience to recover the updated state if needed
instead of different arguments depending on event.